Rename the top level dir of the labs api and pin data set hoster tag
