+++ docker/test/uwsgi/uwsgi-labs-api.ini
@@ -3,7 +3,7 @@ uid = www-data
 gid = www-data
 master = true
 socket = 0.0.0.0:3031
-module = listenbrainz.api.labs.main
+module = listenbrainz.labs_api.labs.main
 callable = app
 chdir = /code/listenbrainz
 processes = 10
