+++ requirements.txt
@@ -23,7 +23,7 @@ oauth2client == 4.1.3
 google-api-python-client==1.9.3
 pika == 0.13.0
 git+https://github.com/metabrainz/brainzutils-python.git@v1.14.1
-git+https://github.com/metabrainz/data-set-hoster.git#egg=datasethoster
+git+https://github.com/metabrainz/data-set-hoster.git@v-2020-08-05.0#egg=datasethoster
 spotipy == 2.12.0
 pyyaml == 5.3.1
 eventlet == 0.25.2
